--- src/test/java/com/cg/onlineadvertisementsystem/OnlineAdvertisementSystemApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.cg.onlineadvertisementsystem;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OnlineAdvertisementSystemApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
